pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TSForecasting_InvoiceGPT"
version = "0.1.0"
description = "InvoiceGPT" 
dependencies = [
    "google-genai>=1.3.0",
    "langchain>=0.3.19",
    "langchain-google-genai>=2.0.11",
    "langgraph>=0.3.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
